+++ src/pluckit/plugins/__init__.py
@@ -1,4 +1,13 @@
 from pluckit.plugins.base import Plugin, PluginRegistry
+from pluckit.plugins.history import Commit, History
 from pluckit.plugins.viewer import AstViewer, Rule, parse_viewer_query
 
-__all__ = ["Plugin", "PluginRegistry", "AstViewer", "Rule", "parse_viewer_query"]
+__all__ = [
+    "Plugin",
+    "PluginRegistry",
+    "AstViewer",
+    "Rule",
+    "parse_viewer_query",
+    "History",
+    "Commit",
+]